refactor: create a component to display container list table Signed-off-by: Philippe Martin <phmartin@redhat.com>
